Returning the live mileage

+++ app/src/main/kotlin/quebec/virtualite/unirider/bluetooth/DeviceInfo.kt
@@ -1,5 +1,5 @@
 package quebec.virtualite.unirider.bluetooth
 
 data class DeviceInfo(
-    val mileage: Int
+    val mileage: Float
 )
